requirements.txt
attrs==20.3.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachelib==0.9.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
cs50==9.2.2
Flask==2.2.1
Flask-Session==0.4.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
jellyfish==0.9.0
Jinja2==3.1.2
lib50==3.0.4
lxml==4.9.1
MarkupSafe==2.1.1
oauthlib==3.2.0
pexpect==4.8.0
ptyprocess==0.7.0
pycparser==2.21
python-dotenv==0.20.0
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
schedule==1.1.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
termcolor==1.1.0
tweepy==4.10.0
twython==3.9.1
tzdata==2022.2
urllib3==1.26.11
Werkzeug==2.2.1
